+++ magda-web-client/src/Components/SearchFacets/FacetBasicBody.js
@@ -99,7 +99,7 @@ class FacetBasicBody extends Component {
     let inactiveOptions = this.props.options.filter(o=>!this.checkActiveOption(o));
     // the option that has the max object.value value, use to calculate volumne indicator
     let maxOptionOptionList = maxBy(this.props.options, o=> +o.hitCount);
-    return (<div className={'facet-body ' + 'facet-' + this.props.title}>
+    return (<div className={`facet-body facet-${this.props.title}`}>
               <div className='clearfix facet-body__header'>
                 <FacetSearchBox renderOption={this.renderOption}
                                 options={this.props.facetSearchResults}