+++ magda-web-client/src/actions/publisherActions.js
@@ -3,7 +3,7 @@
 import fetch from 'isomorphic-fetch'
 import {config} from '../config'
 import {actionTypes} from '../constants/ActionTypes';
-import type {Error } from '../types';
+import type {FetchError } from '../types';
 import type {FacetAction} from '../helpers/datasetSearch';
 
 export function requestPublishers():FacetAction {
@@ -19,7 +19,7 @@ export function receivePublishers(json: Object): FacetAction {
   }
 }
 
-export function requestPublishersError(error: Error): FacetAction {
+export function requestPublishersError(error: FetchError): FacetAction {
   return {
     type: actionTypes.REQUEST_PUBLISHERS_ERROR,
     error,
@@ -39,7 +39,7 @@ export function receivePublisher(json: Object): FacetAction {
   }
 }
 
-export function requestPublisherError(error: Error): FacetAction {
+export function requestPublisherError(error: FetchError): FacetAction {
   return {
     type: actionTypes.REQUEST_PUBLISHER_ERROR,
     error,
@@ -55,13 +55,12 @@ function fetchPublishers(start){
                 if (response.status === 200) {
                     return response.json();
                 }
-                return dispatch(requestPublishersError({title: response.status, detail: response.statusText}))
+                throw(new Error(response.statusText));
             })
             .then(json => {
-                if(!json.error){
-                    return dispatch(receivePublishers(json));
-                }
+                return dispatch(receivePublishers(json));
             })
+            .catch(error => dispatch(requestPublishersError({title: error.name, detail: error.message})))
     }
 }
 
@@ -95,13 +94,12 @@ function fetchPublisher(id){
                 if (response.status === 200) {
                     return response.json()
                 }
-                return dispatch(requestPublisherError({title: response.status, detail: response.statusText}))
+                throw(new Error(response.statusText));
             })
             .then(json => {
-                if(!json.error){
-                    return dispatch(receivePublisher(json));
-                }
+                return dispatch(receivePublisher(json));
             })
+            .catch(error => dispatch(requestPublisherError({title: error.name, detail: error.message})))
     }
 }