+++ CHANGES.md
@@ -13,11 +13,19 @@
 * Updated visualization sleuther to stream file downloads and csv parsing, and relax time field specifications.
 * Added `userId` parameter to `package.json` of `magda-gateway` module
 * Added execution permission to `setup.sh` to solve the issue that `magda-elastic-search` failed to start in minikube
+* Updated format sleuther to be a bit more optimistic in its sleuthing
+* Re-added viz sleuther to default helm config
 * Added index to `"publisher"` field in recordaspects table in order to stop indexer webhook queries taking 10 minutes.
 * Added a CONTRIBUTING.md file
 * Fixed an issue that `Preview Map` doesn't support WFS API
+* Made the indexer listen to delete record events and remove the deleted record from the index
 * Added prettier `pre-commit` hook to make sure consistent code style
 * Formatted existing typescript source code using `prettier`
+* Updated `building-and-running.md`
+* Added preview map support for geojson data type
+* Merged latest changes (commits on or before 1st Feb 2018) from TerrisMap to `magda-preview-map` module
+* Map preview will zoom to dataset (except KML data)
+* Add Google Analytics Tag Manager Code / VWO code to `<head>`
 
 ## 0.0.32